+++ src/app/core/provider/oldvae.service.ts
@@ -1,0 +1,27 @@
+// import { Injectable } from '@angular/core';
+// import { HttpClient, HttpResponse } from '@angular/common/http';
+
+// /**
+//  * Definition of VAE service provider (allows for access to model weights)
+//  * @class
+//  */
+// @Injectable()
+// export class VaeService {
+//     constructor(private http: HttpClient) { }
+
+//     // getVaeWeights() {
+//     //     /**TODO */
+//     // } 
+
+//     // getEncoderMeanWeights(collection: string) {
+//     //     return this.http.get('assets/encodermean' + collection + '.json', {observe: 'response'});
+//     // }
+
+//     // getEncoderLogVarWeights(collection: string) {
+//     //     return this.http.get('assets/encoderlogvar' + collection + '.json', {observe: 'response'});
+//     // }
+
+//     // getDecoderWeights(collection: string) {
+//     //     return this.http.get('assets/decoder' + collection + '.json', {observe: 'response'});
+//     // }
+// }
