+++ src/app/core/provider/patternfinder.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { PatternfinderService } from './patternfinder.service';
+
+describe('PatternfinderService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: PatternfinderService = TestBed.get(PatternfinderService);
+    expect(service).toBeTruthy();
+  });
+});
