+++ src/app/core/provider/vae.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { VaeService } from './vae.service';
+
+describe('VaeService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: VaeService = TestBed.get(VaeService);
+    expect(service).toBeTruthy();
+  });
+});
