+++ src/app/weaver/weaver.component.ts
@@ -8,17 +8,12 @@ import { Draft } from '../core/model/draft';
 import { Render } from '../core/model/render';
 import { Pattern } from '../core/model/pattern';
 import { MatDialog } from "@angular/material/dialog";
-import { InitModal } from '../core/modal/init/init.modal';
 import {Subject} from 'rxjs';
 import { FileService, LoadResponse } from '../core/provider/file.service';
 import { Loom } from '../core/model/loom';
 import * as _ from 'lodash';
-import { PatternFinder } from './tool/patternfinder/patternfinder';
-import { VAE } from './learning/vae';
-import { DraftdetailComponent } from '../mixer/modal/draftdetail/draftdetail.component';
 import { DraftviewerComponent } from '../core/draftviewer/draftviewer.component';
 import {DesignmodesService} from '../core/provider/designmodes.service'
-import { isBuffer } from 'lodash';
 import { SidebarComponent } from '../core/sidebar/sidebar.component';
 import { MaterialsService } from '../core/provider/materials.service';
 import { Cell } from '../core/model/cell';
@@ -87,17 +82,6 @@ export class WeaverComponent implements OnInit {
   */
   collection: string = "";
 
-  /**
-  * Object that holds the machine learning models to generate drafts from a seed
-  * @property {VAE} 
-  */
-  vae: VAE = new VAE();
-   
-  /**
-  * When generativeMode is activated, patternFinder will be run to determine the major patterns of the current draft
-  * @property {PatternFinder}
-  */
-  patternFinder: PatternFinder = new PatternFinder();
 
   /**
   * Number of warps for drafts of collection selected 
@@ -451,59 +435,59 @@ export class WeaverComponent implements OnInit {
     }
     return pattern;
 }
-  /**
-   * Flips the current booleean value of generativeMode.
-  * @extends WeeaveComponent
-  * @param {Event} e
-  * @returns {void}
-  */
- public onGenerativeModeChange(e: any) {
-   console.log('e:', e);
-   this.generativeMode = !this.generativeMode;
-   this.collection = e.collection.toLowerCase().split(' ').join('_');
-   this.warpSize = e.warpSize;
-   this.weftSize = e.weftSize;
-   this.vae.loadModels(this.collection).then(() => {
-    if (this.generativeMode) {
-      this.vae.loadModels(this.collection);
-      let pattern = this.patternFinder.computePatterns(this.loom.threading, this.loom.treadling, this.draft.pattern);
-      var suggestions = [];
-      let draftSeed = this.patternToSize(pattern, this.warpSize, this.weftSize);
-      this.vae.generateFromSeed(draftSeed).then(suggestionsRet => {
-        suggestions = suggestionsRet;
-        console.log('suggestions:', suggestions);
-        for (var i = 0; i < suggestions.length; i++) {
-          let treadlingSuggest = this.patternFinder.getTreadlingFromArr(suggestions[i]);
-          let threadingSuggest = this.patternFinder.getThreadingFromArr(suggestions[i]);
-          let pattern = this.patternFinder.computePatterns(threadingSuggest, treadlingSuggest, suggestions[i])
-          let draft = new Draft({});
-          for (var i = 0; i < pattern.length; i++) {
-            var first = false;
-            if (i != 0) {
-              draft.pattern.push([]);
-            } else {
-              first = true;
-            }
-            for (var j = 0; j < pattern[i].length; j++) {
-              if (first && j == 0) {
-                draft.pattern[i][j] = new Cell(pattern[i][j] == 1 ? true : false);
-              } else {
-                draft.pattern[i].push(new Cell(pattern[i][j] == 1 ? true : false));
-              }
-            }
-          }
-          this.generated_drafts.push(draft);    
-        }
-      });
-    }
-   });
- }
+//   /**
+//    * Flips the current booleean value of generativeMode.
+//   * @extends WeeaveComponent
+//   * @param {Event} e
+//   * @returns {void}
+//   */
+//  public onGenerativeModeChange(e: any) {
+//    console.log('e:', e);
+//    this.generativeMode = !this.generativeMode;
+//    this.collection = e.collection.toLowerCase().split(' ').join('_');
+//    this.warpSize = e.warpSize;
+//    this.weftSize = e.weftSize;
+//    this.vae.loadModels(this.collection).then(() => {
+//     if (this.generativeMode) {
+//       this.vae.loadModels(this.collection);
+//       let pattern = this.patternFinder.computePatterns(this.loom.threading, this.loom.treadling, this.draft.pattern);
+//       var suggestions = [];
+//       let draftSeed = this.patternToSize(pattern, this.warpSize, this.weftSize);
+//       this.vae.generateFromSeed(draftSeed).then(suggestionsRet => {
+//         suggestions = suggestionsRet;
+//         console.log('suggestions:', suggestions);
+//         for (var i = 0; i < suggestions.length; i++) {
+//           let treadlingSuggest = this.patternFinder.getTreadlingFromArr(suggestions[i]);
+//           let threadingSuggest = this.patternFinder.getThreadingFromArr(suggestions[i]);
+//           let pattern = this.patternFinder.computePatterns(threadingSuggest, treadlingSuggest, suggestions[i])
+//           let draft = new Draft({});
+//           for (var i = 0; i < pattern.length; i++) {
+//             var first = false;
+//             if (i != 0) {
+//               draft.pattern.push([]);
+//             } else {
+//               first = true;
+//             }
+//             for (var j = 0; j < pattern[i].length; j++) {
+//               if (first && j == 0) {
+//                 draft.pattern[i][j] = new Cell(pattern[i][j] == 1 ? true : false);
+//               } else {
+//                 draft.pattern[i].push(new Cell(pattern[i][j] == 1 ? true : false));
+//               }
+//             }
+//           }
+//           this.generated_drafts.push(draft);    
+//         }
+//       });
+//     }
+//    });
+//  }
 
- public loadGeneratedDraft(e: any){
-  console.log("running load generated draft!");
+//  public loadGeneratedDraft(e: any){
+//   console.log("running load generated draft!");
   
-  //tell the draft viewer to load this business!
-}
+//   //tell the draft viewer to load this business!
+// }
   
   /**
    * Tell the weave directive to fill selection with pattern.
